Added time units on histogram, recorder, and collector.  #314 Removed unused buffer in shaper.
git-svn-id: https://svn.code.sf.net/p/gridlab-d/code/trunk@988 54b63ff4-7b51-0410-8bda-9fa96797e505

+++ tape/histogram.h
@@ -53,8 +53,8 @@ public:
 	int32 bin_count;
 	double min;
 	double max;
-	int32 sampling_interval;	//	add values into bins
-	int32 counting_interval;	//	dump the bins into a row
+	double sampling_interval;	//	add values into bins
+	double counting_interval;	//	dump the bins into a row
 	int32 limit;			//	number of lines to write
 	
 	FILETYPE type;
